feat: add utilities for graph databases
