+++ packages/kineo/src/adapter.ts
@@ -1,5 +1,6 @@
 import type { IR } from "./ir";
-import type { Schema, SchemaDiff } from "./schema";
+import type { Schema, Node as SchemaNode, SchemaDiff } from "./schema";
+import Model from "./model";
 
 /**
  * Parameters. Used for passing parameters into queries.
@@ -51,6 +52,18 @@ export type OptPromise<T> = T | Promise<T>;
  */
 export type Adapter = {
   /**
+   * Constructor of a model class. Not necessary if no extension is used.
+   */
+  Model?: {
+    new <S extends Schema, N extends SchemaNode, A extends Adapter>(
+      label: string,
+      schema: S,
+      node: N,
+      adapter: A
+    ): Model<S, N, A>;
+  };
+
+  /**
    * Gets schema from database.
    */
   getSchema(): OptPromise<Schema>;
@@ -84,7 +97,7 @@ export type Adapter = {
    */
   status(
     migrations: string[],
-    hashes: string[],
+    hashes: string[]
   ): OptPromise<Array<"deployed" | "pending">>;
 
   /**
@@ -120,7 +133,7 @@ export const isNode = (v: unknown): v is Node => v instanceof Node;
  */
 export const getScalar = <T extends Scalar>(
   rec: QueryRecord | undefined,
-  key: string | number,
+  key: string | number
 ): T | undefined => {
   const v = rec?.get?.(key);
   return isNode(v) ? undefined : (v as T | undefined);
@@ -136,7 +149,7 @@ export const getScalar = <T extends Scalar>(
 export const getNodeProp = <T>(
   rec: QueryRecord | undefined,
   idx: number | string,
-  prop: string,
+  prop: string
 ): T | undefined => {
   const v = rec?.get?.(idx);
   return isNode(v) ? (v.properties[prop] as T | undefined) : undefined;
@@ -174,7 +187,7 @@ export class Node {
     identity: number | bigint,
     labels: string[],
     properties: Params,
-    elementId: string,
+    elementId: string
   ) {
     this.identity = identity;
     this.labels = labels;
