+++ packages/kineo/src/compilers/cypher.ts
@@ -1,18 +1,19 @@
 import type { Compiler } from "../adapter";
 import type {
-  IRConnect,
   IRDelete,
   IRCreate,
   IRMatch,
   IRMerge,
-  IRRelationQuery,
   IRStatement,
   IRWhereNode,
+} from "../ir";
+import { IRConnect, IRRelationQuery } from "../graph/ir";
+import {
   IRGetNodeLabels,
   IRGetNodeProperties,
   IRGetRelationshipTypes,
   IRGetRelationshipProperties,
-} from "../ir";
+} from "../adapters/neo4j/ir";
 
 /**
  * Compiles an intermediate representation to Cypher.
@@ -260,7 +261,7 @@ function compileRelationQuery(stmt: IRRelationQuery, idx: number) {
 
   // MATCH the relationship
   commandLines.push(
-    `MATCH (${from.alias})-[:${rel}]->(${alias}:${stmt.label})`,
+    `MATCH (${from.alias})-[:${rel}]->(${alias}:${stmt.label})`
   );
   if (relWhere.clause) {
     commandLines.push(`WHERE ${relWhere.clause}`);
@@ -287,7 +288,7 @@ function compileRelationQuery(stmt: IRRelationQuery, idx: number) {
 function compileWhere(
   node: IRWhereNode | undefined,
   alias: string,
-  idx: number,
+  idx: number
 ): {
   clauses: string[];
   params: Record<string, unknown>;
@@ -350,7 +351,7 @@ function compileWhere(
       inner.push(
         n.OR.map(walk)
           .map((s) => `(${s})`)
-          .join(" OR "),
+          .join(" OR ")
       );
     if (n.NOT) inner.push(`NOT (${walk(n.NOT)})`);
 
@@ -372,7 +373,7 @@ function compileWhere(
 function compileMatchObject(
   where: Record<string, unknown>,
   alias: string,
-  idx: number,
+  idx: number
 ) {
   const clauses: string[] = [];
   const params: Record<string, unknown> = {};
